+++ src/frame.c
@@ -64,8 +64,8 @@ void page_init(size_t memory, size_t table)
 	numpages = ROUND_UP(memory, PAGE_SIZE) / PAGE_SIZE;
 
 
-	int r = ROUND_UP((numpages * sizeof(struct page)), PAGE_SIZE) / PAGE_SIZE;
-	int i = 0;
+	size_t r = ROUND_UP((numpages * sizeof(struct page)), PAGE_SIZE) / PAGE_SIZE;
+	size_t i = 0;
 
 	/* Handle cases where total number of physical pages will not fit inside
 	 * of the previously mapped virtual area (0-4MB)... I.e. this will
@@ -77,7 +77,7 @@ void page_init(size_t memory, size_t table)
 	assert((table + 3 * PAGE_SIZE) < INITIAL_TOP);
 	if ((r * PAGE_SIZE + table) > V2P(INITIAL_TOP)) {
 		/* Calculate how many 2MB pages are needed and map them in */
-		int req = ROUND_UP((r*PAGE_SIZE + table), 0x200000);
+		size_t req = ROUND_UP((r*PAGE_SIZE + table), 0x200000);
 		mmu[0] = table;
 		mmu[1] = table + (PAGE_SIZE*1);
 		mmu[2] = table + (PAGE_SIZE*2);
